backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
httpx
Pillow
google-genai
huggingface_hub
